AAS_ApplyFriction: full long double + sqrtl to match original x87 80-bit precision
comments — i.e. live x87 stack slots — but the operands and sqrt() call were
(or worse, 32-bit float for the intermediate vel[i]*vel[i] sums) and then
widened only at assignment.

Promote every operand explicitly to long double, switch sqrt -> sqrtl, and
narrow the stores back to float on assignment to vel[]. This matches the
identical pattern already applied to VectorLength and the JUMP projection
math, and brings the per-frame friction integration inside
AAS_ClientMovementPrediction in line with the original x87 80-bit precision.

no regressions, q2dm1 and q2ctf8 still each missing one JUMP record (the
residual traces to AAS_ClientMovementPrediction's larger physics loop, not
to friction).

Source-commit: 82c869dffc42c3389f2df85af918e2581459f285

+++ botlib/botlib.c
@@ -9788,18 +9788,18 @@ void __cdecl AAS_ApplyFriction(vec3_t vel, float friction, float stopspeed, floa
   long double v5; // st6
   long double v6; // st6
 
-  v4 = sqrt(vel[0] * vel[0] + vel[1] * vel[1]);
-  if ( v4 != 0.0 )
+  v4 = sqrtl((long double)vel[0] * (long double)vel[0] + (long double)vel[1] * (long double)vel[1]);
+  if ( v4 != 0.0L )
   {
-    if ( v4 >= stopspeed )
+    if ( v4 >= (long double)stopspeed )
       v5 = v4;
     else
-      v5 = stopspeed;
-    v6 = v4 - v5 * friction * frametime;
-    if ( v6 < 0.0 )
-      v6 = 0.0;
-    vel[0] = v6 / v4 * vel[0];
-    vel[1] = v6 / v4 * vel[1];
+      v5 = (long double)stopspeed;
+    v6 = v4 - v5 * (long double)friction * (long double)frametime;
+    if ( v6 < 0.0L )
+      v6 = 0.0L;
+    vel[0] = (float)(v6 / v4 * (long double)vel[0]);
+    vel[1] = (float)(v6 / v4 * (long double)vel[1]);
   }
 }
